pyproject.toml
[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core >= 3.4, < 4"]

[project]
authors = [
  {name = "Jacqueline Ryan", email = "Jacqueline.Ryan@jpl.caltech.edu"},
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License", # Adjust based on your actual license
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Scientific/Engineering :: Image Processing",
]
dependencies = [
  "pvl == 1.3.2",
  "numpy >= 1.20",
  "pillow >= 9.0",
]
description = "Utility for converting .VIC/.IMG images to compressed image formats."
license = {file = "LICENSE"}
name = "vic2png"
readme = "README.md"
requires-python = ">=3.8"
version = "0.3.0"

[project.optional-dependencies]
dev = [
  "black",
  "flit",
  "mypy",
  "pyright",
  "pytest",
  "pytest-cov",
  "ruff",
]

[tool.flit.module]
name = "vic2png"

[tool.pytest.ini_options]
addopts = "--cov --cov-report=term-missing --cov-report=xml --cov-report=html --import-mode=importlib"
testpaths = ["test"]

[project.scripts]
vic2png = "vic2png.cli:main"

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.ruff]
ignore = ["I001"]
line-length = 88
select = ["E", "F", "I", "N", "W", "B", "Q"]
target-version = "py38"

[tool.mypy]
disallow_untyped_defs = true
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true

[tool.coverage.run]
branch = true
source = ["vic2png"]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "if self.debug:",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "pass",
  "raise ImportError",
]
ignore_errors = true
omit = [
  "tests/*",
  "setup.py",
]
